Update wifi_manager.py
On the ESP32  after being connected to one wifi and that wifi gets unavaliable for any reason...its not possible to scan a new network to connect to without disconnect from the one that was connected. Because the folowing error would appear: "wifi:sta_scan: STA is connecting, scan are not allowed!"
The solution is to disconnect from the previous before trying to search for a new one.